Introduce daily bug stats chart
This commit introduces daily bug stats graphs. To do this, this commit
also introduces nvd3 js library for enabling to use lineWithFocusChart.
It is not same style chart of the main pages ones. Because
stackedAreaWithFocusChart is a little bit buggy. So it sometimes shows
wrong charts. And I also think lineWithFocusChart is better than the
stackedArea chart because it is simple enough to understand the trend.

Change-Id: Iacf22cddbd676d1447cfd9ae8ba1e7ce387720f6